Cleanu up folder structure, move gnn networks into networks folder
