+++ bark_ml/library_wrappers/lib_tf_agents/networks/BUILD
@@ -1,0 +1,11 @@
+py_library(
+    name = "networks",
+    srcs = ["__init__.py",
+            "gnn_wrapper.py",
+            "gnn_actor_network.py",
+            "gnn_critic_network.py"],
+    data = ["@bark_project//bark/python_wrapper:core.so"],
+    imports = ["../external/bark_project/bark/python_wrapper/"],
+    deps = [],
+    visibility = ["//visibility:public"],
+)
