+++ bark_ml/library_wrappers/lib_tf_agents/networks/__init__.py
@@ -1,0 +1,3 @@
+from bark_ml.library_wrappers.lib_tf_agents.networks.gnn_actor_network import GNNActorNetwork
+from bark_ml.library_wrappers.lib_tf_agents.networks.gnn_critic_network import GNNCriticNetwork
+from bark_ml.library_wrappers.lib_tf_agents.networks.gnn_wrapper import GNNWrapper
