+++ examples/BUILD
@@ -65,23 +65,6 @@ py_test(
   deps = [
     "//bark_ml/environments:single_agent_runtime",
     "//bark_ml/environments:blueprints",
-    "//bark_ml/library_wrappers/lib_tf2_gnn:lib_tf2_gnn",
-    "//bark_ml/library_wrappers/lib_tf_agents/agents:agents",
-    "//bark_ml/library_wrappers/lib_tf_agents/runners:runners",
-  ]
-)
-
-py_test(
-  name = "hyperparameter_optimization",
-  srcs = ["hyperparameter_optimization.py"],
-  data = ["@bark_project//bark/python_wrapper:core.so",
-          "//bark_ml:generate_core",
-          "//examples/example_params:example_params"],
-  imports = ["../bark_ml/python_wrapper/"],
-  deps = [
-    "//bark_ml/environments:single_agent_runtime",
-    "//bark_ml/environments:blueprints",
-    "//bark_ml/library_wrappers/lib_tf2_gnn:lib_tf2_gnn",
     "//bark_ml/library_wrappers/lib_tf_agents/agents:agents",
     "//bark_ml/library_wrappers/lib_tf_agents/runners:runners",
   ]
